Revision includes three changes:
Added double-precision floating point operations to 15411 library. This will be broken on systems with 4-byte pointers, but this is kind of an unofficial library from the perspective of mainstream C0 anyway so I'm not worried about. Fixed a library bug with passing/returning NULL to native functions.


Corresponding bump of the C0VM version number because I added functions to a library.

Modified cc0 to return success if asked for version with -V or help with -h, since it will have done what it's supposed to do. This will facilitate tools that try running C0 to make sure they have the right path.

git-svn-id: https://svn.concert.cs.cmu.edu/c0@513 59102d7d-8565-4dee-862f-f79749aaa1f2